refactor(worker): `vc-processors-v2` -> `vc-processors`
