--- venus-worker/vc-processors-v2/Cargo.toml
@@ -1,66 +1,0 @@
-[package]
-name = "vc-processors-v2"
-version = "0.2.0"
-edition = "2021"
-license = "MIT"
-keywords = ["filecoin", "venus", "venus-cluster", "venus-worker"]
-description = "provide core types & builtin processors for venus-worker"
-homepage = "https://github.com/ipfs-force-community/venus-cluster"
-repository = "https://github.com/ipfs-force-community/venus-cluster"
-
-[features]
-default = []
-numa = ["lazy_static"]
-
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-anyhow = "1"
-fnv = "1.0.7"
-tokio = { version = "1.26", features = ["macros", "rt"] }
-thiserror = "1"
-tokio-transports = { version = "0.1", default-features = false, features = [
-    "serded-json",
-    "tracing",
-    "rw-pipe",
-] }
-futures-util = { version = "0.3", default-features = false, features = [
-    "sink",
-] }
-tokio-util = { version = "0.7", features = ["time"] }
-pin-project = "1"
-lru = "0.10"
-indexmap = "1.9.2"
-rand = "0.8"
-tower = { version = "0.4", features = [
-    "util",
-    "balance",
-    "ready-cache",
-    "discover",
-    "limit",
-] }
-bincode = "1.3"
-hex = "0.4"
-twox-hash = "1.6.3"
-lazy_static = { version = "1.4", optional = true }
-tracing = "0.1"
-tracing-opentelemetry = { version = "0.18", default-features = false }
-opentelemetry_api = { version = "0.18", default-features = false, features = [
-    "trace",
-] }
-
-[target.'cfg(target_os = "linux")'.dependencies]
-cgroups-rs = "0.3"
-
-[dev-dependencies]
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-tracing-test = { version = "0.2", features = ["no-env-filter"] }
-pretty_assertions = "1.3"
-tempfile = "3"
-tokio = { version = "1", features = ["full"] }
-tokio-test = "0.4"
-tower-test = "0.4"
-
-[[example]]
-name = "sqr"
-required-features = []
