--- venus-worker/vc-processors-v2/src/sys/mod.rs
@@ -1,6 +1,0 @@
-//! This module provides some operating system level APIs or bindings of native libraries
-//!
-
-pub mod cgroup;
-#[cfg(feature = "numa")]
-pub mod numa;
