+++ venus-worker/vc-processors/Cargo.toml
@@ -1,90 +1,66 @@
 [package]
 name = "vc-processors"
-version = "0.2.1"
+version = "0.2.0"
 edition = "2021"
 license = "MIT"
-keywords = [
-    "filecoin",
-    "venus",
-    "venus-cluster",
-    "venus-worker",
-]
+keywords = ["filecoin", "venus", "venus-cluster", "venus-worker"]
 description = "provide core types & builtin processors for venus-worker"
 homepage = "https://github.com/ipfs-force-community/venus-cluster"
 repository = "https://github.com/ipfs-force-community/venus-cluster"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[features]
+default = []
+numa = ["lazy_static"]
 
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-tracing = "0.1"
 anyhow = "1"
-crossbeam-channel = { version = "0.5", optional = true }
-memmap = { version = "0.7", optional = true }
-uuid = { version = "1", features = ["v4", "fast-rng"] }
-forest_address = { version = "0.3", optional = true }
-base64-serde = { version = "0.7", optional = true }
-base64 = { version = "0.21", optional = true }
-lazy_static = "1.4"
-reqwest = { version = "0.11", features = ["blocking"], optional = true}
 fnv = "1.0.7"
+tokio = { version = "1.26", features = ["macros", "rt"] }
+thiserror = "1"
+tokio-transports = { version = "0.1", default-features = false, features = [
+    "serded-json",
+    "tracing",
+    "rw-pipe",
+] }
+futures-util = { version = "0.3", default-features = false, features = [
+    "sink",
+] }
+tokio-util = { version = "0.7", features = ["time"] }
+pin-project = "1"
+lru = "0.10"
+indexmap = "1.9.2"
+rand = "0.8"
+tower = { version = "0.4", features = [
+    "util",
+    "balance",
+    "ready-cache",
+    "discover",
+    "limit",
+] }
+bincode = "1.3"
+hex = "0.4"
+twox-hash = "1.6.3"
+lazy_static = { version = "1.4", optional = true }
+tracing = "0.1"
+tracing-opentelemetry = { version = "0.18", default-features = false }
+opentelemetry_api = { version = "0.18", default-features = false, features = [
+    "trace",
+] }
 
 [target.'cfg(target_os = "linux")'.dependencies]
 cgroups-rs = "0.3"
 
-[dependencies.filecoin-proofs]
-version = "12.0"
-default-features = false
-features = ["opencl", "multicore-sdr"]
-optional = true
-
-[dependencies.filecoin-proofs-api]
-version = "12.0"
-default-features = false
-features = ["opencl", "multicore-sdr"]
-optional = true
-
-[dependencies.storage-proofs-core]
-version = "12.0"
-optional = true
-
-[dependencies.storage-proofs-porep]
-version = "12.0"
-optional = true
-
-
 [dev-dependencies]
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-tracing-test = "0.2"
-rand = "0.8"
+tracing-test = { version = "0.2", features = ["no-env-filter"] }
 pretty_assertions = "1.3"
-hex = "0.4"
 tempfile = "3"
-
-[features]
-default = ["ext-producer", "builtin"]
-ext-producer = ["crossbeam-channel"]
-numa = []
-b64serde = ["base64", "base64-serde" ]
-fil-proofs = ["memmap", "b64serde", "forest_address", "filecoin-proofs", "filecoin-proofs-api", "storage-proofs-core", "storage-proofs-porep"]
-fil-proofs-cuda = ["fil-proofs", "filecoin-proofs/cuda", "filecoin-proofs-api/cuda"]
-builtin-tasks = ["fil-proofs"]
-builtin-processors = ["builtin-tasks", "reqwest"]
-builtin = ["builtin-tasks", "builtin-processors"]
+tokio = { version = "1", features = ["full"] }
+tokio-test = "0.4"
+tower-test = "0.4"
 
 [[example]]
-name = "sqr-consumer"
+name = "sqr"
 required-features = []
-
-[[example]]
-name = "sqr-producer"
-required-features = ["ext-producer"]
-
-[[example]]
-name = "customize-proc"
-required-features = ["ext-producer", "builtin-tasks"]
-
-[[example]]
-name = "builtin-proc"
-required-features = ["ext-producer", "builtin"]
