--- venus-worker/vc-processors/src/builtin/mod.rs
@@ -1,9 +1,0 @@
-//! Built-in tasks & processors
-
-#[cfg(feature = "builtin-tasks")]
-#[allow(missing_docs)]
-pub mod tasks;
-
-#[cfg(feature = "builtin-processors")]
-#[allow(missing_docs)]
-pub mod processors;
